documentation: remove links to element/plugin docs that don't exist
If an element or plugin is not documented, we'd just link to a page that
doesn't exist. There is an explanatory note at the top of the page that
explains that this means the element or plugin in question is not
documented, but people seem to regularly miss that note or still find
it confusing to run into 404s.

This commit adds a little perl helper script that removes the hyperlinks
to pages that don't exist. This also makes it much easier to see which
elements are actually documented or not.