new-release: error out if the RELEASE file was not generated for some reason
This may happen for example when there's a syntax error in the XML
(e.g. forgot to escape some chars in the features list) or the
build system in www failed for some reason (e.g. because it has
been a while since autogen was run last and automake has been
upgraded since).