download: Add ABI compatibility notes for Windows
People often have misconceptions regarding this, such as:

https://gitlab.freedesktop.org/gstreamer/gstreamer/issues/386
https://gitlab.freedesktop.org/gstreamer/gstreamer/issues/387

People also need to know what CRT(s) each set of binary releases uses.
For example, there are different CRTs for debug vs release, and the
CRT can be linked statically or dynamically.

This time use tables so that we can control the styling and because
CMark doesn't support tables.