www: Update layout/sidebar to look a bit more modern
Centers the main container and limits it to 1400px width, making it easier to look at on a widescreen display.

Makes the sidebar look like buttons, not just text with a background.
Changes it so that the entire button there is a link, not just the text - feels much better to use :)
Adds a bit of padding for better UX, plus adjusts colors, borders and spacing etc.

Part-of: <https://gitlab.freedesktop.org/gstreamer/www/-/merge_requests/105>